Refactor models into components
All model imports are now separate from AnimatedCamera
