+++ src/components/AnimatedCamera.js
@@ -1,17 +1,18 @@
-/*
-GLTF imports with the help of: https://gltf.pmnd.rs/ https://github.com/pmndrs/gltfjsx
-*/
-
 import * as THREE from 'three';
 import React, { useRef, useState } from 'react';
 import { useGLTF, useAnimations, PerspectiveCamera } from '@react-three/drei';
 import { useFrame } from '@react-three/fiber';
 import { useLocation } from 'wouter';
-import UselessCubes from './UselessCubesGroup';
-import ComputerModel from './Computer';
-import ArcadeModel from './Arcade';
-import BowserModel from './Bowser';
-import Mario from './Mario';
+import UselessCubesComponent from './models/UselessCubesGroup';
+import HackerNewsModel from './models/Computer';
+import FakeNewsModel from './models/Arcade';
+import BowserModel from './models/Bowser';
+import MarioModel from './models/Mario';
+import {
+  DoomGuyModel as ThreeArenaModel,
+  DeloreanModel as Electric,
+  NookBagModel as NookIncModel,
+} from './models/ModelGroup';
 
 export default function Models({ scroll, ...props }) {
   const time = useRef(0);
@@ -31,7 +32,7 @@ export default function Models({ scroll, ...props }) {
   const animEndQuaternion = new THREE.Quaternion(0.0928, -0.3936, 0.04, 0.9136);
   const thirdPhaseQuart = new THREE.Quaternion(0, 0.7, 0, 0.7122);
 
-  const { nodes, materials, animations } = useGLTF('models/modelGroup.glb');
+  const { animations } = useGLTF('models/modelGroup.glb');
   const { actions, mixer } = useAnimations(animations, group);
 
   const [cameraReady, setCameraReady] = useState(false);
@@ -49,10 +50,6 @@ export default function Models({ scroll, ...props }) {
   const scrollElement = document.querySelector('.scroll');
   const projectButtons = document.querySelectorAll('.project-buttons');
   const devNav = document.querySelector('#dev-nav');
-
-  const extras = {
-    'material-envMapIntensity': 0.2,
-  };
 
   if (!toggle) {
     if (locationId !== '') {
@@ -278,8 +275,7 @@ export default function Models({ scroll, ...props }) {
         position={[0.06, 4.04, 0.35]}
         scale={[0.25, 0.25, 0.25]}
       >
-        {/****  START OF ALL MODEL GROUPS  ****/}
-        {/* New Bowser & Mario */}
+        {/****  START OF ALL MODELS  ****/}
         <group
           onClick={(e) => {
             e.stopPropagation();
@@ -290,11 +286,9 @@ export default function Models({ scroll, ...props }) {
             );
           }}
         >
-          <Mario />
+          <MarioModel />
           <BowserModel />
         </group>
-
-        {/* hacker news model */}
         <group>
           <group
             position={window.innerWidth < 800 ? [42, -2, 1] : [42, -2, 5]}
@@ -305,11 +299,9 @@ export default function Models({ scroll, ...props }) {
               setTimeout(() => setLocation(toggle ? '/hacker-news' : '/'), 500);
             }}
           >
-            <ComputerModel />
+            <HackerNewsModel />
           </group>
         </group>
-
-        {/* three arena model */}
         <group
           onClick={(e) => {
             e.stopPropagation();
@@ -317,88 +309,8 @@ export default function Models({ scroll, ...props }) {
             setTimeout(() => setLocation(toggle ? '/three-arena' : '/'), 500);
           }}
         >
-          <group
-            position={[33.65, window.innerHeight < 800 ? 10.5 : 12, -17.34]}
-            rotation={[-Math.PI, 0.6, -Math.PI]}
-            scale={[0.2, 0.2, 0.2]}
-          >
-            <primitive object={nodes._rootJoint} />
-            <mesh
-              name="ThreeArena"
-              geometry={nodes.rock_GEO_Rock_MAT_0.geometry}
-              material={materials.Rock_MAT}
-            />
-            <mesh
-              name="ThreeArena"
-              geometry={nodes.sides_GEO_Side_MAT_0.geometry}
-              material={materials.Side_MAT}
-            />
-            <mesh
-              name="ThreeArena"
-              geometry={nodes.ground_GEO_treeline_MAT_0.geometry}
-              material={materials.treeline_MAT}
-            />
-            <mesh
-              name="ThreeArena"
-              geometry={nodes.rocks_GEO_Rock_MAT_0.geometry}
-              material={materials.Rock_MAT_0}
-            />
-            <mesh
-              name="ThreeArena"
-              geometry={nodes.shadow_GEO_Shadow_MAT_0.geometry}
-              material={materials.Shadow_MAT}
-            />
-            <skinnedMesh
-              name="ThreeArena"
-              geometry={nodes.sword.geometry}
-              material={materials.Doom_MAT_0}
-              skeleton={nodes.sword.skeleton}
-            />
-            <skinnedMesh
-              name="ThreeArena"
-              geometry={nodes.flame.geometry}
-              material={materials.Doom_MAT_1}
-              skeleton={nodes.flame.skeleton}
-            />
-            <skinnedMesh
-              name="ThreeArena"
-              geometry={nodes.shotgun.geometry}
-              material={nodes.shotgun.material}
-              skeleton={nodes.shotgun.skeleton}
-            />
-            <skinnedMesh
-              name="ThreeArena"
-              geometry={nodes.fists.geometry}
-              material={nodes.fists.material}
-              skeleton={nodes.fists.skeleton}
-            />
-            <skinnedMesh
-              name="ThreeArena"
-              geometry={nodes.body.geometry}
-              material={nodes.body.material}
-              skeleton={nodes.body.skeleton}
-            />
-            <skinnedMesh
-              name="ThreeArena"
-              geometry={nodes.arms.geometry}
-              material={materials.Doom_MAT_3}
-              skeleton={nodes.arms.skeleton}
-            />
-            <skinnedMesh
-              name="ThreeArena"
-              geometry={nodes.blade.geometry}
-              material={materials.Doom_MAT_2}
-              skeleton={nodes.blade.skeleton}
-            />
-            <skinnedMesh
-              name="ThreeArena"
-              geometry={nodes.head.geometry}
-              material={materials.Doom_MAT}
-              skeleton={nodes.head.skeleton}
-            />
-          </group>
+          <ThreeArenaModel />
         </group>
-        {/* electric model */}
         <group
           onClick={(e) => {
             e.stopPropagation();
@@ -406,32 +318,14 @@ export default function Models({ scroll, ...props }) {
             setTimeout(() => setLocation(toggle ? '/electric' : '/'), 500);
           }}
         >
-          <group
-            position={[14.71, window.innerWidth < 800 ? 19.21 : 23.21, -32.71]}
-            scale={[3, 3, 3]}
-            rotation={[-1.44, -0.15, -0.67]}
-          >
-            <mesh
-              name="Electric"
-              geometry={nodes.Delorean.geometry}
-              material={materials.M_Delorean}
-              {...extras}
-            />
-            <mesh
-              name="Electric"
-              geometry={nodes.DeloreanShock.geometry}
-              material={materials.M_Shock}
-            />
-          </group>
+          <Electric />
         </group>
-        {/* useless cubes model */}
-        <UselessCubes
+        <UselessCubesComponent
           onClickEvent={() => {
             setToggle(!toggle);
             setTimeout(() => setLocation(toggle ? '/useless-web' : '/'), 500);
           }}
         />
-        {/* nook inc model */}
         <group
           onClick={(e) => {
             e.stopPropagation();
@@ -439,166 +333,8 @@ export default function Models({ scroll, ...props }) {
             setTimeout(() => setLocation(toggle ? '/nook-inc' : '/'), 500);
           }}
         >
-          <group
-            position={[
-              window.innerWidth < 800 ? -20 : -36.34,
-              window.innerWidth < 800 ? 18 : 28,
-              -12,
-            ]}
-            rotation={[-Math.PI, -1.05, -Math.PI]}
-            scale={
-              window.innerWidth < 800
-                ? [0.055, 0.055, 0.055]
-                : [0.04, 0.04, 0.04]
-            }
-          >
-            <group
-              position={[-5.59, 134.86, -11.56]}
-              rotation={[-Math.PI / 2, 0, 1.31]}
-              scale={[100, 100, 100]}
-            >
-              <mesh
-                name="NookInc"
-                geometry={nodes.bell_Bag_bag_Mat00_0.geometry}
-                material={materials.bag_Mat00}
-              />
-            </group>
-            <group
-              position={[-3.66, 177.74, -18.8]}
-              rotation={[-1.76, 0.03, 1.32]}
-              scale={[45.01, 44.27, 61.54]}
-            >
-              <mesh
-                name="NookInc"
-                geometry={nodes.bell_Ribbon001_ribbon_Mat00_0.geometry}
-                material={nodes.bell_Ribbon001_ribbon_Mat00_0.material}
-              />
-            </group>
-            <group
-              position={[59.12, 145.46, 56.44]}
-              rotation={[-3.07, 0.39, -2.68]}
-              scale={[100, 100, 100]}
-            >
-              <mesh
-                name="NookInc"
-                geometry={nodes.bell_Ribbon002_ribbon_Mat00_0.geometry}
-                material={nodes.bell_Ribbon002_ribbon_Mat00_0.material}
-              />
-            </group>
-            <group
-              position={[-91.11, 133.84, -34.65]}
-              rotation={[-0.07, -0.39, 0.46]}
-              scale={[100, 100, 100]}
-            >
-              <mesh
-                name="NookInc"
-                geometry={nodes.bell_Ribbon_ribbon_Mat00_0.geometry}
-                material={nodes.bell_Ribbon_ribbon_Mat00_0.material}
-              />
-            </group>
-            <group
-              position={[84.4, 86.6, 117.71]}
-              rotation={[2.81, 0.65, -1.55]}
-              scale={[100, 100, 100]}
-            >
-              <mesh
-                name="NookInc"
-                geometry={nodes.bell_Coin_bell_Mat00_0.geometry}
-                material={nodes.bell_Coin_bell_Mat00_0.material}
-              />
-            </group>
-            <group
-              position={[-138.53, 304.91, -75.78]}
-              rotation={[2.95, 0.51, 2.72]}
-              scale={[100, 100, 100]}
-            >
-              <mesh
-                name="NookInc"
-                geometry={nodes.bell_Coin001_bell_Mat00_0.geometry}
-                material={nodes.bell_Coin001_bell_Mat00_0.material}
-              />
-            </group>
-            <group
-              position={[46.72, 428.62, -33.36]}
-              rotation={[2.81, 0.65, -1.55]}
-              scale={[100, 100, 100]}
-            >
-              <mesh
-                name="NookInc"
-                geometry={nodes.bell_Coin002_bell_Mat00_0.geometry}
-                material={nodes.bell_Coin002_bell_Mat00_0.material}
-              />
-            </group>
-            <group
-              position={[122.47, 351.52, -116.6]}
-              rotation={[-2.33, 0.76, 2.3]}
-              scale={[100, 100, 100]}
-            >
-              <mesh
-                name="NookInc"
-                geometry={nodes.bell_Coin003_bell_Mat00_0.geometry}
-                material={nodes.bell_Coin003_bell_Mat00_0.material}
-              />
-            </group>
-            <group
-              position={[-185.29, 143.48, 29.52]}
-              rotation={[3.02, -0.57, 0.98]}
-              scale={[100, 100, 100]}
-            >
-              <mesh
-                name="NookInc"
-                geometry={nodes.bell_Coin004_bell_Mat00_0.geometry}
-                material={nodes.bell_Coin004_bell_Mat00_0.material}
-              />
-            </group>
-            <group
-              position={[-91.38, 419.22, -30.81]}
-              rotation={[2.04, -0.01, 3.08]}
-              scale={[100, 100, 100]}
-            >
-              <mesh
-                name="NookInc"
-                geometry={nodes.bell_Coin005_bell_Mat00_0.geometry}
-                material={nodes.bell_Coin005_bell_Mat00_0.material}
-              />
-            </group>
-            <group
-              position={[-119.72, 205.86, -180.75]}
-              rotation={[-2.44, -0.51, 1.44]}
-              scale={[100, 100, 100]}
-            >
-              <mesh
-                name="NookInc"
-                geometry={nodes.bell_Coin007_bell_Mat00_0.geometry}
-                material={nodes.bell_Coin007_bell_Mat00_0.material}
-              />
-            </group>
-            <group
-              position={[142.76, 253.41, -10.37]}
-              rotation={[-1.44, 0.8, -0.7]}
-              scale={[100, 100, 100]}
-            >
-              <mesh
-                name="NookInc"
-                geometry={nodes.bell_Coin008_bell_Mat00_0.geometry}
-                material={nodes.bell_Coin008_bell_Mat00_0.material}
-              />
-            </group>
-            <group
-              position={[193.25, 271.61, 97.33]}
-              rotation={[-2.86, 0.39, 2.39]}
-              scale={[100, 100, 100]}
-            >
-              <mesh
-                name="NookInc"
-                geometry={nodes.bell_Coin009_bell_Mat00_0.geometry}
-                material={nodes.bell_Coin009_bell_Mat00_0.material}
-              />
-            </group>
-          </group>
+          <NookIncModel />
         </group>
-
-        {/* Fake News model */}
         <group>
           <group
             visible={
@@ -608,7 +344,7 @@ export default function Models({ scroll, ...props }) {
             rotation={[0.3, -1.1, 0.2]}
             scale={[0.4, 0.4, 0.4]}
           >
-            <ArcadeModel
+            <FakeNewsModel
               onClick={(e) => {
                 e.stopPropagation();
                 setToggle(!toggle);
